LsMenuBar & LsMenuItem
This is a replacement for the built in MenuBars and ToolBars.
The classes are able to behave like a MenuBar or a ToolBar with
grouping and an overflowmenu.

--- --- --- --- --- --- --- ---

Title:    LsMenuBar
Version:  0.1

Author:   Patric Lee Atrott
Web:      www.atrott.net

### ### ### ### ### ### ### ###
Notes:
--- --- --- --- --- --- --- ---

LsMenuBar & LsMenuItem and pending classes, structures, etc.
are not finished yet and can be very buggy from time to time.

This class is able to use MBS Plugins.
It is highly recommended to use MBS Plugins.

### ### ### ### ### ### ### ###
Contributors:
--- --- --- --- --- --- --- ---

The XOJO Community

  // Cocoa_Mouseclick Code by: Jason K
  //
https://forum.xojo.com/12285-move-mouse-cursor-and-click-by-code-for-mac
  // 25 May 2014

  // Windows_Mouseclick Code from: Michel B - XOJO Forum
  //
https://forum.xojo.com/12259-move-mouse-position-click-by-code-windows
  // 25 May 2014

Icons by Yusuke Kamiyamane.
Fugue Icons
  (C) 2013 Yusuke Kamiyamane. All rights reserved.
  These icons are licensed under a Creative Commons
  Attribution 3.0 License.
  <http://creativecommons.org/licenses/by/3.0/>
  If you can't or don't want to provide attribution, please
  purchase a royalty-free license.
  <http://p.yusukekamiyamane.com/>

### ### ### ### ### ### ### ###
License:
--- --- --- --- --- --- --- ---

CC BY SA 4.0

http://creativecommons.org/licenses/by-sa/4.0/
http://creativecommons.org/licenses/by-sa/4.0/legalcode

You are free to:

    Share — copy and redistribute the material in any medium or format
    Adapt — remix, transform, and build upon the material
    for any purpose, even commercially.

    The licensor cannot revoke these freedoms as long as you follow the
license terms.

Under the following terms:

    Attribution — You must give appropriate credit, provide a link to
the license, and indicate if changes were made. You may do so in any
reasonable manner, but not in any way that suggests the licensor
endorses you or your use.

    ShareAlike — If you remix, transform, or build upon the material,
you must distribute your contributions under the same license as the
original.

    No additional restrictions — You may not apply legal terms or
technological measures that legally restrict others from doing anything
the license permits.

Notices:

    You do not have to comply with the license for elements of the
material in the public domain or where your use is permitted by an
applicable exception or limitation.
    No warranties are given. The license may not give you all of the
permissions necessary for your intended use. For example, other rights
such as publicity, privacy, or moral rights may limit how you use the
material.