Realbasic / Xojo - This is a replacement for the built in MenuBars and ToolBars.
LsMenuBar

This is a replacement for the built in MenuBars and ToolBars. The
classes are able to behave like a MenuBar or a ToolBar with grouping
and an overflowmenu.

Version: 0.1

### ### ### ### ### ### ###

ToDos:

--- --- --- --- --- --- --- --- ---

- keybindings -> keystrokes for easy use of menuitems

- toggle groups

### ### ### ### ### ### ###

Notes:

--- --- --- --- --- --- --- --- ---

LsMenuBar & LsMenuItem and pending classes, structures, etc.

are not finished yet and can be very buggy from time to time.

This class is able to use MBS Plugins.
It is highly recommended to use MBS Plugins.